requirements.txt
zeroconf==0.132.2
requests==2.32.3
psutil==6.0.0
PyYAML==6.0.2

